% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_divergent}
\alias{aefcolor_divergent}
\title{AEF info color palettes: divergent}
\format{
\subsection{\code{aefcolor_divergent}}{

A character vector of length 4
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_divergent
}
\description{
Divergent color palette from all the palettes set by the French
news agency AEF info graphic team.
}
\keyword{datasets}
