% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_extended}
\alias{aefcolor_extended}
\title{AEF info color palettes: extended}
\format{
\subsection{\code{aefcolor_extended}}{

A character vector of length 20
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_extended
}
\description{
Extended color palette from all the palettes set by the French news agency AEF info graphic team.
}
\keyword{datasets}
