% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_gradientblue}
\alias{aefcolor_gradientblue}
\title{AEF info color palettes: gradient blue}
\format{
\subsection{\code{aefcolor_gradientblue}}{

A character vector of length 5
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_gradientblue
}
\description{
Blue gradient color palette from all the palettes set by the
French news agency AEF info graphic team.
}
\keyword{datasets}
