% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_gradientgreen}
\alias{aefcolor_gradientgreen}
\title{AEF info color palettes: gradient green}
\format{
\subsection{\code{aefcolor_gradientgreen}}{

A character vector of length 5
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_gradientgreen
}
\description{
Green gradient color palette from all the palettes set by the
French news agency AEF info graphic team.
}
\keyword{datasets}
