% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_gradientorange}
\alias{aefcolor_gradientorange}
\title{AEF info color palettes: gradient orange}
\format{
\subsection{\code{aefcolor_gradientorange}}{

A character vector of length 5
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_gradientorange
}
\description{
Orange gradient color palette from all the palettes set by the French news agency AEF info graphic team.
}
\keyword{datasets}
