% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_gradientred}
\alias{aefcolor_gradientred}
\title{AEF info color palettes: gradient red}
\format{
\subsection{\code{aefcolor_gradientred}}{

A character vector of length 5
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_gradientred
}
\description{
Red gradient color palette from all the palettes set by the French news agency AEF info graphic team.
}
\keyword{datasets}
