% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_gradientviolet}
\alias{aefcolor_gradientviolet}
\title{AEF info color palettes: gradient violet}
\format{
\subsection{\code{aefcolor_gradientviolet}}{

A character vector of length 5
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_gradientviolet
}
\description{
Violet gradient color palette from all the palettes set by the French news agency AEF info graphic team.
}
\keyword{datasets}
