% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_main}
\alias{aefcolor_main}
\title{AEF info color palettes: main colors}
\format{
\subsection{\code{aefcolor_main}}{

A character vector of length 2
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_main
}
\description{
Main color palette from all the palettes set by the French news agency AEF info graphic team.
}
\keyword{datasets}
