% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_recommended}
\alias{aefcolor_recommended}
\title{AEF info color palettes: recommended}
\format{
\subsection{\code{aefcolor_recommended}}{

A character vector of length 6
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_recommended
}
\description{
Recommended color palette from all the palettes set by the
French news agency AEF info graphic team.
}
\keyword{datasets}
