% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aefcolor_secondary}
\alias{aefcolor_secondary}
\title{AEF info color palettes: secondary}
\format{
\subsection{\code{aefcolor_secondary}}{

A character vector of length 5
}
}
\source{
\url{https://www.aefinfo.fr/ressource/charte/}
}
\usage{
aefcolor_secondary
}
\description{
Secondary color palette from all the palettes set by the French news agency AEF info graphic team.
}
\keyword{datasets}
