% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fb_messages_sample_df}
\alias{fb_messages_sample_df}
\title{Number of messages sent by Facebook Messenger "hiho" conversation members}
\format{
\subsection{\code{fb_messages_sample_df}}{

A data frame with 9 rows and 2 columns:
\describe{
\item{sender_name}{Name of the sender}
\item{messages}{Total messages sent}
}
}
}
\usage{
fb_messages_sample_df
}
\description{
Sample data frame to build examples within the aefplot package.
}
\keyword{datasets}
